collect bigint helpers in one place
