+++ src/lib/encoding.ts
@@ -1,3 +1,4 @@
+import { bytesToBigInt, changeBase } from '../js_crypto/bigint-helpers.js';
 import { fieldEncodings } from '../provable/base58.js';
 import { Field } from '../snarky.js';
 
@@ -144,72 +145,6 @@ function toFieldsBijective(bytes: Uint8Array, p: bigint) {
   return fields;
 }
 
-// various helpers
-
-function changeBase(digits: bigint[], base: bigint, newBase: bigint) {
-  // 1. accumulate digits into one gigantic bigint `x`
-  let x = fromBase(digits, base);
-  // 2. compute new digits from `x`
-  let newDigits = toBase(x, newBase);
-  return newDigits;
-}
-
-// NOTE: toBase / fromBase are so complicated for performance reasons
-
-function fromBase(digits: bigint[], base: bigint) {
-  // compute powers base, base^2, base^4, ..., base^(2^k)
-  // with largest k s.t. n = 2^k < digits.length
-  let basePowers = [];
-  for (let power = base, n = 1; n < digits.length; power **= 2n, n *= 2) {
-    basePowers.push(power);
-  }
-  let k = basePowers.length;
-  // pad digits array with zeros s.t. digits.length === 2^k
-  digits = digits.concat(Array(2 ** k - digits.length).fill(0n));
-  // accumulate [x0, x1, x2, x3, ...] -> [x0 + base*x1, x2 + base*x3, ...] -> [x0 + base*x1 + base^2*(x2 + base*x3=, ...] -> ...
-  // until we end up with a single element
-  for (let i = 0; i < k; i++) {
-    let newDigits = Array(digits.length >> 1);
-    let basePower = basePowers[i];
-    for (let j = 0; j < newDigits.length; j++) {
-      newDigits[j] = digits[2 * j] + basePower * digits[2 * j + 1];
-    }
-    digits = newDigits;
-  }
-  console.assert(digits.length === 1);
-  let [digit] = digits;
-  return digit;
-}
-
-function toBase(x: bigint, base: bigint) {
-  // compute powers base, base^2, base^4, ..., base^(2^k)
-  // with largest k s.t. base^(2^k) < x
-  let basePowers = [];
-  for (let power = base; power < x; power **= 2n) {
-    basePowers.push(power);
-  }
-  let digits = [x]; // single digit w.r.t base^(2^(k+1))
-  // successively split digits w.r.t. base^(2^j) into digits w.r.t. base^(2^(j-1))
-  // until we arrive at digits w.r.t. base
-  let k = basePowers.length;
-  for (let i = 0; i < k; i++) {
-    let newDigits = Array(2 * digits.length);
-    let basePower = basePowers[k - 1 - i];
-    for (let j = 0; j < digits.length; j++) {
-      let x = digits[j];
-      let high = x / basePower;
-      newDigits[2 * j + 1] = high;
-      newDigits[2 * j] = x - high * basePower;
-    }
-    digits = newDigits;
-  }
-  // pop "leading" zero digits
-  while (digits[digits.length - 1] === 0n) {
-    digits.pop();
-  }
-  return digits;
-}
-
 // a constant field is internally represented as {value: [0, Uint8Array(32)]}
 function bytesOfConstantField(field: Field): Uint8Array {
   let value = (field as any).value;
@@ -226,16 +161,6 @@ function bigIntToField(x: bigint) {
   let field = Field(1);
   (field as any).value = [0, bigIntToBytes(x, 32)];
   return field;
-}
-
-function bytesToBigInt(bytes: Uint8Array) {
-  let x = 0n;
-  let bitPosition = 0n;
-  for (let byte of bytes) {
-    x += BigInt(byte) << bitPosition;
-    bitPosition += 8n;
-  }
-  return x;
 }
 
 function bigIntToBytes(x: bigint, length: number) {
